+++ Tests/AIBatteryCoreTests/Views/ActivityChartIsEmptyTests.swift
@@ -80,6 +80,7 @@ struct ActivityChartIsEmptyTests {
             rateLimits: nil,
             rateLimitSource: nil,
             standardLimits: nil,
+            rateLimitPercentConfirmed: true,
             firstSessionDate: nil,
             totalSessions: 0,
             totalMessages: 0,